Adding cast to int in auto LB comparison in hopes it would work

+++ WEB-INF/src/ece1779/servlets/LoadBalancerLibrary.java
@@ -457,11 +457,11 @@ public class LoadBalancerLibrary {
 				ratioShrink = 2;
 			}
 			
-			if(avgLoad > cpuThresholdGrowing)
+			if((int)avgLoad > cpuThresholdGrowing)
 			{
 				increaseWorkerPoolSize(currPoolSize * ratioExpand, (AWSCredentials)servletContext.getAttribute("AWSCredentials"));
 			}
-			else if(avgLoad < cpuThresholdShrinking)
+			else if((int)avgLoad < cpuThresholdShrinking)
 			{
 				decreaseWorkerPoolSize(currPoolSize / ratioShrink, (AWSCredentials)servletContext.getAttribute("AWSCredentials"));
 			}
